Fix: Resolve lint errors with TS implementation
\### Rationale
Best practices within Typescript are checked with a linter and this
commit changes the implementation to be in spec. Large recommendation
is to never have an explicit `any` type since it defeats the purpose
of Typescript's type-checking protections.

Digging into `@typescript-eslint` the provided types definitions
determined what values are provided back from Eslint.