fix: expectExpect rule to handle more use cases
\### Rationale
Due to the lack of type-checking and not utilizing recursion to review
the function tree, there were some missing cases where the rule would
fail. Added more test cases which highlight code sequences that do
fail on the old implementation.

Also, made strict adjustments to jest threshold related to this file's
error handling. So if this changes, it will have to be determined
of a better implementation.