fix: abstract rule imports & use getters
\### Rationale
In order to facilitate better testing of the correct imported rules,
the rules imports need to be abstracted rather than using a global var
which is not abstractable and controllable for invalid case checking.

Also converted to getter functions to ensure imports are low
computation during import but when accessed the computation occurs.