fix-test: replicate prod pkg install in e2e test
\### Rationale
NPM takes a shortcut when running an install with just `file:..` and
only creates a symlink. Make the e2e test as real as possible with a
built & packed production version tarball that then is installed
directly.  Peer Dependencies will be resolved, and other dependencies
will also have to be resolved which will identify any other install
errors that occured in `v1.2.6`.