fix(rule-metadata): fixed issue with generation of manual url with codebase use of version
Preventative measure to ensure URLs are managed better in case version entries change. This was
primarily noticed due to the webpack build of the url generated with 0.0.0-semantically-versioned
entry was now in the url during testing, weird case but not caught. This change intends to make it
more realistic but it will still be v0.0.0 otherwise default to master.