style(lint-conf): revamp eslint matching methodology to ignore-all-except
Found out that `--ext` only applies if a directory is passed to `eslint@v8` so pre-commit hook was
linting ignored files. Revamped so `--ext` is not used but `.eslintignore` is used instead.  Huge
performance gain as well since folders that were not supposed to be evaluated were being evaluated
anyway! Runtime is about 20sec.